+++ Plugins/BrawlInstaller Plugins/Install/Install Franchise Icon.py
@@ -1,5 +1,4 @@
 __author__ = "Squidgy"
-__version__ = "1.7.0"
 
 from BrawlInstallerLib import *
 from InstallLib import *
@@ -14,13 +13,15 @@ def main():
 			if not Directory.Exists(MainForm.BuildPath + '/pf/'):
 				BrawlAPI.ShowMessage("Build path does not appear to be valid. Please change your build path by going to 'Tools > Settings' and modifying the 'Default Build Path' field.\n\nYour build path should contain a folder named 'pf' within it.", "Invalid Build Path")
 				return
-			createLogFile()
-
 			# Get user settings
 			if File.Exists(MainForm.BuildPath + '/settings.ini'):
 				settings = getSettings()
 			else:
 				settings = initialSetup()
+			if not settings:
+				return
+
+			createLogFile()
 			# If temporary directory already exists, delete it to prevent duplicate files
 			if Directory.Exists(AppPath + '/temp'):
 				Directory.Delete(AppPath + '/temp', 1)
